requirements.txt
flask==2.2.5 
flask-sqlalchemy==2.5.1 
flask-login==0.6.3 
flask-migrate==4.1.0 
flask-mail==0.10.0 
python-dotenv==1.0.1 
gunicorn==23.0.0 
sqlalchemy==1.4.54 
werkzeug==2.2.3 
psycopg2-binary==2.9.9 
requests>=2.31  
# OIDC Authentication
oauthlib==3.2.2 
requests-oauthlib==1.3.1 
PyJWT==2.8.0 
cryptography==41.0.5 
pyOpenSSL==23.2.0 
flask-apscheduler
pytz>=2021.1
tzdata>=2021.1
Flask-Migrate>=3.1.0
Flask-SQLAlchemy>=2.5.1
alembic>=1.7.5

psycopg2-binary>=2.9.1  # For PostgreSQL

# REST API Dependencies
flask-jwt-extended==4.6.0  # JWT authentication for API
flask-cors==4.0.0  # CORS support for React Native
marshmallow==3.21.0  # Serialization/deserialization
marshmallow-sqlalchemy==1.0.0  # SQLAlchemy integration
flask-restx==1.3.0  # REST API with Swagger documentation

# Investment integrations
multitasking==0.0.11  # Required by yfinance, 0.0.12+ requires Python 3.9+
yfinance==0.2.18  # Yahoo Finance stock data (0.2.28+ requires Python 3.9+)

# 29a41de6a866d56c36aba5159f45257c
